pyproject.toml
[tool.poetry]
name = "store-warehouse"
version = "0.1.0"
description = ""
authors = ["Daniil <danokp@mail.ru>"]
readme = "README.md"
packages = [{include = "store_warehouse"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
python-dotenv = "^1.0.0"
dj-database-url = "^2.0.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.2.7"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
